+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Added
 
+- Added an option to hide subprogram sockets by default.
+- Added default values for Script subprogram inputs.
 
 ### Fixed